requirements.txt
flask==2.0.1
zyphra
gunicorn
python-dotenv
requests 